Improve Continuous Intergration
Implement CI also for macOS. Test merge requests on our macOS 10.13 test system.
The pipline now consists of two steps with three different jobs. First the compliance
with the code formating rules is checked. If this step succeed the code is build and
tested on Linux and macOS. This step runs simultaneous on both platforms.