Fix AnalysisTree and AnalysisTreeQA installation
Use commit hash of a tag instead of tag name to define the needed git version.
The hash is needed to check the already downloaded and build version.
Without the hash after each run of CMake the libraries are rebuild.
Using the hash only rebuild the libraries if there is a change of the hash.