Automatically build doxygen documentation
For each commit to the branch nightly_master a build of the doxygen
documentation is triggered. The branch is updated once per night if there
are changes to the master branch. The branch reflects the state of the
master branch at midnight.
The documentation is build using a docker container and the result is
uploaded to the gitlab pages web page at
http://computing.gitpages.cbm.gsi.de/cbmroot/