Add custom streamer for CbmPsdDigi (#1814)
The entry for the CbmPsdDigi in the file DataLinkdef.h was wrong, such that
for the CbmPsdDigi a streamer was generated which does not support automatic
schema evolution. To overcome the problem a custom streamer with explicit
schema evolution was added. This streamer allows to read also CbmPsdDigi
Version 3 data from existing files. Refs #1814