Allow as input ROOT file from ZIP archives
Since the size of a single PLUTO file is very small the preferred storage
solution for them is to store a large number of them in a archive.
ROOT is able to handle ROOT files stored in a ZIP archive so the file
handling in the class was adapted for the ZIP file usage.