Convert old to new tof adressing scheme
Add code which automatically converts the old tof address scheme to
the new one for the the CbmTofDigi data class.
Since for some of the varibales less bits than before are
available a check was added which catches such cases and stop the
execution with an error message in such cases.
This is unfortunately not yet possible for CbmTofPoint and CbmTofHit.
For both data classes the address is a member of the base class and
it was not possible to change the value of the base class member in
the same way. When reading any old versions of both classes the
macro execution is stopped immediately with an error message.