Adapt build system to new FairSoft and FairRoot versions
Add support for C++17
Check for c++ version c++11 or above (11, 14 or 17). With this change it
becomes possible to compile CbmRoot with C++17.

Add missing library.
The FairMQ state machine is now in an extra library which has to be added
for linking when using FairSoft nov20. Test if the library exist and create
the list of dependencies accordingly.

Add missing include directory.
VMC is not any longer included in ROOT but a standealone package, so the
include directory has to be added such that the header files are found.

Add gtest as an optional external package.
With the latest version of FairSoft googletest was removed due to a
recommendation of the gtest developers. Since gtest is no longer part of
FairSoft we have to provide an own version which is compiled as external
project.

Add proper dependency to external gtest project
This dependency is needed to compile the test only if the gtest library is
already available.

Fix compilation error with FairRoot v18.4.2 on macosx
In FairRoot v18.4.2 some linker settings were removed from the CMake setup.
This breaks the compilation of CbmRoot. Put back the needed seetings for the
linker on macosx in our build system.

Fix NicaFemto installation