Add function which checks if an input is a ROOT file
The function Cbm::File::IsRootFile checks if a file with the given name
exists and if this is confirmed it checks if the file is either a ROOT
file or a ROOT file stored in a ZIP file. ROOT is able to handle ROOT
files stored in a ZIP archive.
The function returns true when the file is a ROOT file  and false in
any other case.