Fix install target
When installing CbmRoot using FairRoot v18.4 and FairSoft there was a problem
with the location of header files when executing macros. Some of the header
files couldn't be found which for some of them worked with printing warning
or error messages only. For other header files the macros simply crashed with
error messages about not found header files. It is unclear if this is a
FairRoot or Root issue but I would guess it is a Root problem.
The issue was solved by simply removing the path from the header file name when
generating the Root dictionary. In this case it is enough to specify the
directory where all header files are installed when running Root using the
environment variable ROOT_INCLUDE_PATH.

In some of the build system files missing directories where added such that
the dictionaries could be produced properly.