Check if the last line is terminated by a <newline> character
According to the Posix standard a line of a text is defined as follows:

3.206 Line
A sequence of zero or more non- <newline> characters plus a terminating
<newline> character.

The test checks if this is valid for all newly added or changed text files.

I stumbeled over the issue when adding a new header file the list of header
files used to generate a ROOT dictionary.
The previous last file of the list did not have a proper file ending and
attaching the new header file results in a ROOT dictionary source code with
the following line

'#endif#ifdef'

which could be properly parsed and compiled by the compiler but results in an
error at run time when loading the library.

To avoid such problems in future the test was added.