Fix usage of OpenMP on macosx
On macosx one needs a openmp installation which does not come with the system,
at least the omp.h header file is missing. brew has a package libomp which
installs the header and a library. Both are found by the cmake module
FindOpenMP and properly used during the compilation and linking.
In themoment there is still a runtime error when executing the L1 tracking
with OpenMP support on macosx, so switch off OpenMP for the time being.
The problem is investigated already. refs #2028.

Cleanup OpenMP handling in the build system. Move check for OpenMP to top
level. Remove OpenMP usage from all libs were it is not needed.