Prepare TOF code for new geometry tof_v20a
Don't use tof digi parameter file any longer in the digitization step.
The needed information is meanwhile taken from the TGeoManager.

Initialize tof geometry parameters in the simple clusterizer from the
TGeoManager, as it is already done in the event clusterizer and the digitizer.

Add a special case in the TofGeoManager for the tof_v20a geometry.
The geometry has to be treated as v14a.