Fix code owner functionality
There was a problem with the script such that in some cases to many code
owners were informed. When in between the creation of the MR and the
runtime of the script the master branch of the official repository was
updated the code owners of the MR changes as well as the code owners of
the updated code from the official repository wer informed.
To avoid sending information for code which is already implemented in the
official repository the local working copy is rebased before extracting the
changed files.

Fix user name of Sergey Gorbunov in CODEOWNERS file. Sergey has two GitLab
accounts and the wrong was chosen.