Fix CI pipeline after GitLab upgrade
Fix the stage for the rebase check.
After the upgrade it lloks like that the pipeline doesn't run any longer in
the context of the forked project but in context of the master project such
that the project path wasn't correct any longer. This allows now to compare
upstream/master to HEAD.