Rerun failed tests a second time
For the continuous tests rerun failed tests a second time. This should reduce
the number of failing test pipelines.
In most cases of failing tests the Geant3 simulations crash due to the well
known memory error. If these tests are restarted they normally work without
problem.
Since the feature to restart tests is only available with cmake 3.17 change
the scripts of the CI pipeline to use at least this version.