Updates cmake tags to include geometry and parameter files for tof_v20b.
The commit add the geometry and parameter files for tof_v20b from the
external parameter and geometry repositories.
The commit as well changes the tof version in the json config file to tof_v20b.
The geometry tof_v20b fixes an issue with tof_v20a and tof_v16e found on
18.04.21.
The number of channels per module and some module types were incorrect.
For consistency the existing geometry tof_v20a wasn't changed but a new fixed
one tof_v20b was created.
The macro used for the generation of the tof_v20 geometries was commited
with 0a48b9fc.
The complete disscusions about the new geometry are at
https://git.cbm.gsi.de/computing/cbmroot/-/merge_requests/314
https://git.cbm.gsi.de/CbmSoft/cbmroot_geometry/-/merge_requests/83