Include script which allows a costum setup file to be created.
Script which asks a series of questions which detectors to include to allow automatic generation of a user specified setup script.
The scripts allows only default geometries as a feature. Additionally, allows the user to shift the downstream detectors
to any desired location with correct placements prompts. The script doesn't not modify existing geometries but
creates new geometries with the auto tag which are overwritten on each subsequent run.

Although linking of parameter files has been included. This script was created for the management of geometries and should be treated with caution for use in simulation.
The script creates new macros using the prenom "costum" from the run_transport, run_digi and run_reco digis.