Fix config scripts (again)
The previous fix did not work for macosx 11.3. Somehow on this system the old
FindGEANT3.cmake file was used.
Probably this is not a problem with the macosx version but with the usage of
a case insensitive file system. To avoid the problem remove the old and
obsolete file.
Also do some fixes to properly initialise some cmake variables.