Fix a compilation error introduced with MR 323
An external packages is only updated if there is a change in the commit hash
of the external package.
Unfortunately there was no change of the hash value in MR 323 but only a patch
setp was added.
The introduced patch is such not applied since no action is taken at all.
This commit fixes the problem by changing the commit hash by one.
The new commit only adds code which isn't compiled in our setup such is safe.