Add tests on macosx 1015 and 11
Add tests for  macosx 10.15 and 11 on our new runners.
Retire old test on macosx 10.13.

Define number of CPUs in runner configuration.

Move the definition from the global gitlab CI configuration to the local
runner configurations.
Defining the number of CPUs in the runner configuration allows to adjust this
value to the actual underlying hardware.