Include the TRD2D into the scheme plus some minor improvements
For the current situation with the TRD1D and TRD2D configs trying to create and access in principle the same "TrdDigi" branch a small workaround in the CbmRecoUnpack was required. Such that the first Config does not loose the correct pointer when the seconds overwrites the branch in the FairRootManager.