Adapt code to new FairModule function
From FairRoot version v18.4.0 on the function CheckIfSensitive() of the class
FairModule is deprecated and replaced by the function IsSensitive(). The old
function is still available but usage results is a compilation error.
To be able to use FairRoot v18.2.1 and newer versions both functions are
implemented in the CbmRoot code where the old function simply forwards to the
new one. If there is no need to support older FairRoot versions the old
function can be simply removed from the source code.