Fix a performance issue with KFParticle
It was found that switching form FairSoft jun19p2 to nov20 or apr21 has a
dramatic influence on the performance of the KFParticleFinder. The runtime
for on of the tests increased by roughly a factor 16.
https://redmine.cbm.gsi.de/issues/2220
After long investigations it turned out that the compilation flags for the
package were inherited from the ROOT and the CbmRoot project which switched
of the performance flag "-O" when using FairSoft nov20 or apr21.
The merge request fixes the issue by defining the CMAKE_BUILD_TYPE explicitely
to be RelWithDebugInfo.
The merge request also updates the VC version from 1.4.1 to 1.4.2. refs #2220