Correct order of link directories
Somehow (at least on some systems) the link order was such that the
first directory was $SIMPATH/lib. This was even the case when the order was
defined differently in the CMakeLists.txt. The problem isn't yet understood
completely.

This results in the problem that libVc.a was taken from $SIMPATH/lib
were it is installed by ROOT/FairSoft. As long as the version provided with
CbmRoot and the version provided by ROOT were the same the problem wasn't
seen. When updating our internal Vc version this results in a creash when
loading libryries due to a version mismatch between the library version and
the header file version.