Fix the size of the TS MSB storage for the dupli filter
Bug 1: Size was too small which led to some digis with TS_MSB above 65555 escaping the duplicates filter.
No effect on the digi themselves (time is computed from the original 64b/32b value).

In the legacy case we can stay with 32b (instead of the original 64b or the old 16b) as the field in the
raw message is 29b, while in the current version we need 64b to account for potential TS_MSB cycles and
large Timeslice start offset.