Remove ROOT data types from CbmStsDigi
Also remove the dependency to RtypesCore.h.
This is the first step to an additional ROOT free version of the CbmStsDigi
requested for the new algorithm code.
The change of the data types has no influence on the streamer. It is still
possible to read previous versions from file even without increasing the
version number.