Use different test cases with CMake variable
Before the change it was necessary to define an environment variable to switch
between different test suites like for example the Experimental or Nightly
tests. Now the test suites are chosen by the CMake command line parameter
CBM_TEST_MODEL. These are Experimental, Nightly and Weekly which differ by the
number of executed tests.