STS geometries version v21f and v21g
v21f: STS geometry is similar to the version v21b. The target position considered from -4 cm.
The geometry can be used with the beampipe with diameter 40 mm upstream side and 104 mm downstream side.
The geometry has beampipe flange on the back-side of the wall.

v21g: STS geometry is similar to the version v21e. The target position is at -40 cm considering global origin
from the center of magnet. The STS local coordinate is from the last station.
The geometry can be used with the beampipe with diameter 40 mm upstream side and 104 mm downstream side.

In both geometry, the STS box and the beam pipe flange is made of Carbon fiber: The material has 70% carbon fiber with
density 1.75 g/cm3 and epoxy resin (30%) with density 1.2 g/cm3. The material is defined in media.geo file as
STS_CarbonFiber_21.