Fix script
The script contained two problems. The first one was a typo in the shebang.
The second problem was in the grep command. If the full path includes by
chance the string "so" or "dylib" all files of the directory were found
accidentally. Requiring "so" or "dylib" to be at the end of the line fixes
the error.