Improve run script
Add functionality to check if the macro dif finish successful.
The decision can be done on base of the return value as well as on base of
a string printed at the end of the macro.
Don't check the return value yet since oOn some system there are errors when
destructing the root session. Use only the printout from the macro to decide
if the macro finished successful.