STS geometry version v21[h-i]
STS v21h: The geometry has the target positioned at -4 cm. It can be used with the beampipe 40mm diameter upstream side
and 104mm diameter downstream side. It has a beampipe flange on the back-side of the STS wall.
In addition, the geometry has a sensor thickness 320 um which is similar thickness of the real silicon sensors purchased from the Hamamatsu Photonics
for the use in the real experiment.
Furthermore,the mico-cable is defined as a single volume attached to the silicon sensors. Micro-cable thickness (320 um) is choosen and
the material Manganese is taken based on the calculation done by taking the current microcable design, thickness and its material.

STS v21i: The geometry has the target positioned at -44 cm from the center of the magnet. It can be used with the beampipe 40mm diameter upstream side
and 104mm diameter downstream side. It doesn't have a beam-pipe flange on the back-side of the wall.
In addition, the geometry has a sensor thickness 320 um which is similar thickness of the real silicon sensors purchased from the Hamamatsu >
for the use in the real experiment.
Furthermore,the mico-cable is defined as a single volume attached to the silicon sensors. Micro-cable thickness (320 um) is choosen and
the material Manganese is taken based on the calculation done by taking the current microcable design, thickness and its material.