Add new test stage in CTest configuration
Installl CbmRoot and remove part of the source directory

Add code which tests if the installtion of CbmRoot works and if the installed
version is functional. Test the "make install" target and run some macros from
teh installation directory.

To be able to test the CbmRoot installation it is necessary that no header
files are accessible in the source directory since this is the fallback for
ROOT when compiling a macro. To achieve this the directories containing any
headers are removed from the source directory.
This is a dangerous operation when executed in a source directory used for
active development. The installation test should only be used when running
the CI/CD workflow.