Fix CI
Allow to overwrite existing files. This is needed for the CI since the rerun
of failing transport macros wil fail otherwise in any case due to the already
existing output file.
The default setting in the transport macros is to not overwrite existing
files. The CI uses a paramter to change the default.