Fix generation of RICH geometry file
Use the correct TGeo unit system.
Use TVolume->Export to store the RICH volume in the ROOT file.
Add the correct placement matrix according to the shift of the coordinate
system.
Also create the file needed for the media check.
Create all needed materials prior to the import from the gdml file such that
they are properly exported to the ROOT file