Don't transport more events as available in input
In #2421 a crash was reported when the number of events requested to be
transported is larger than the number of events in the input.
Add protection against such a crash in the transport class. The number of
events to be transported is set to the minimum number of events in any of the
connected inputs.