Install MQ startup scripts
Create MQ startup scripts to be installed already at the configuration stage
and store the in the build tree. The difference between the scripts used in
the build tree and the scripts finally installed are the hardcoded directory
information inside. In the scripts to be used inside the build directory this
links to the source and the build directories whereas it links to the
installation directory in case of the scripts which are installed.
Copy the scripts during installation to the correct directory.
Install the MQ executables in the same directory structure as in the build
directory.