Update sis100_electron and all sis100_hadron setups
In the sis100_electron setup the mvd geometry mvd_v20a_tr is used which moves
the mvd to the correct position after the origin was shifted. (#2411)
In the three sis100_hadron setups the mvd was removed completely. (#2386)

This MR brings in the changes from the geometry repository for merge request
https://git.cbm.gsi.de/CbmSoft/cbmroot_geometry/-/merge_requests/143