Fix compiler warnings when using Clang
The KF and L1 packages set special compiler flags. Some of the warnings are
not present when using Clang such that a CMake check was implemented which
test the existense of those. The check simply compiles some C++ code with the
compiler flag added to the compile command. Unfortunately Clang only prints a
warning when a warning flag doesn't exist but successfuly finishes the
compilation such that the check suceeds. This results in many warnings during
compilation of L1.