Switch VM runtime
Use qemu instead of VirtualBox to execute the virtual machines.
The advantage is a much faster and reliable startup of the VMs.
With VirtualBox the image needs to be copied as well there were sometimes
problems with the ssh connection to a running machine.
With qemu the copy of the image isn't needed and no ssh problems were
seen yet.