Fix compiler error on several systems
Depending on the compiler the STL library support or does not support the
usage of std::execution. Tets in the build system if the feature is available
and switch off the usage if not supported.

Remove default(none) from OpenMP parameters since it creates problems with
older OpenMP versions. In the current code according to Jan the statement
isn't needed.