Fix Continuous Integration
The external flesnet requires at least CMake 3.14.0.
Check already in our CMakeLists.txt if this version is available, otherwise
the test is done when building flesnet and the error is hidden in some log
file.
Use CMake from the FairSoft installation on the CI runner (lxir116 )where the
CMake version is to old.