Bugfix material budget Sts macro
Previously, radiation thickness of materials that transversed a cut boundary
were ignored, leading to an underestimation of radiation length. Fixes this.

Allow non equidistant station cuts, such as is needed for the new 5+3 station
STS geometries.

Fixes MC generator to start before the first station of the STS. Sets the MC
generate to the target position for angle generator.