clang-format: Fix header ordering.
Fix a bug in clang-format that causes STL headers that start with a 't'
being grouped with Root headers. (Affects for example <type_traits>.)
This is caused by regex matching in clang-format being case insensitve by default. The option
for case sensitive matching is only introduced with clang-format 12. So
as a workaround for clang-format 11 also match a '.h'-suffix for the
filenames in these categories. This ensures seperation between Root headers
and STL headers.