Fix schema evolution
With commit 892f01a1c63610b02ebb5b7187db3c9aa0278b71 from 23.03.22 the layout
of the data class CbmTsEventHeader was changed by adding a new data member. At
that time it was forgotten to also increase the version number. This commit
increases the version number and corrects the issue.