Implement proper dependencies
With the change to the CMake targets the dependency tree was broken. The
generation of the ROOT dictionary is now a target on its own which did not
depend on anything except the header files needed for the specific library.
Some of these header files include other header files from the external
libraries. The problem was that now it became possible that the dictionary
generation started before the external header files were available.

Add explicit dependencies from these dictionaries to the external libraries.