Move rich local reconstruction code
There were rich ring finders in the L1 code base which were compiled and linked
into libL1. With the change in the build system this results in a circular
dependency between the L1 and CbmRichReco libraries which was resolved by
removing the classes from compilation.
With this commit the classes are moved from the L1 code base into the rich
reconstruction code base.