Prepare for FairRoot v18.8
Use the same target names as provided by FairRoot with version v18.8.

Add the new target FairRoot::Online which is a new library conatining the
online related code which was previously part of the Base library.
In case of older FairRoot versions the target is only an alias for the target
FairRoot::Base such that the same dependencies already work with FairRoot
v18.6 which is our current default version.

Add the new target FairRoot::Online in all CMakeLists.txt where it is needed.

The FindFairRoot macro is also used when compiling NicaFemto. Since there is a
problem with the target definitions in NicaFemto the targets are not defined
when compiling NicaFemto.