Prepare for FairRoot v18.8
Exclude tutorial/TaskToAlgo from build. The code needs to use the boost
serializer for the MCPoint classes. The necessary code was removed from the
FairMCPoint base class such that the compilation currently fails.
The feature can be brought back adding another header file but before adding
it I would like to discuss if this is needed at all.