Fix problem with empty PsdHits
Create and store matches for the psd detector system.

The ideal event builder relies on existing match objects. If the matches are
not available the detector system is excluded from the event building such
that at later stages no digis for this detector system ar present and no hits
are produced.

Fixes #2665.