Use latest version of external repository
Switch to new branch newFieldMaps which adds the various versions of the new
fieldmaps. The switch is needed to be able to remove the maps if they are
updated or become obsolete. If these maps would be included in the main branch
tehre wouldn't be a possibility to remove them at some point.