Fix warnings about deprecated code
With FairRoot v18.8 the functions GetOutputFile() and SetOutputFile() are
deprecated instead of using a file sink. Correct user code to create the
file sink correctly or to extract the file name from the file sink.
Comparisons of results before and after the change showed that the results
are absolutely identical.